# MM-46105: Simplify posts batch retrieval query